Add function and event listener to create grid
I also added styling css file to make sure divs appeared in a square
grid using flexbox. The triggering event is DOM load and initally
creates a 16x16 grid. The function I made is also able to be used
to replace the current square grid with others of different sizes.